requirements.txt
# --- Core ---
python-dotenv==1.0.1
pydantic>=2.0,<3.0 # на 21.05.2025 в проекте установлена: pydantic==2.11.4, pydantic_core==2.33.2
tqdm>=4.66.3

# --- LangChain Ecosystem ---
langchain==0.1.16
langchain-core==0.1.51
langchain-community==0.0.34
langchain-openai==0.1.3
chromadb==0.4.24

# --- Local LLM Inference ---
llama-cpp-python==0.2.73
sentence-transformers==2.6.1
transformers==4.40.0
accelerate==0.23.0
torch>=2.2.2,<2.3.0 --extra-index-url https://download.pytorch.org/whl/cu121 # было, что тоже рабочий вариант: torch==2.2.2+cu121 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
scipy==1.13.0
auto-gptq==0.5.1
peft==0.6.2
sentencepiece
datasets
numpy==1.26.4
prophet==1.1.6

# --- PDF/Text Processing ---
pypdf==3.17.4
unstructured==0.12.5

# --- Web Agents & APIs ---
beautifulsoup4==4.12.3
httpx==0.27.0
requests>=2.32.2
newsapi-python==0.2.7

# --- Gradio UI ---
gradio==4.29.0
gradio_client==0.16.1

# --- Dev & Debug ---
ipykernel==6.29.4

